move init and load logic into /pkd/private/node file
